[software] enable floating downscaling

+++ src/aliceVision/image/resampling.hpp
@@ -47,7 +47,7 @@ namespace image {
   }
 
   template <typename SamplerType, typename Image>
-  void downscaleImage(const Image& src, Image& out, int downscale)
+  void downscaleImage(const Image& src, Image& out, float downscale)
   {
       const int new_width = src.Width() / downscale;
       const int new_height = src.Height() / downscale;
@@ -67,7 +67,7 @@ namespace image {
   }
 
   template <typename ImageType>
-  void downscaleImageInplace(ImageType& inout, int downscale)
+  void downscaleImageInplace(ImageType& inout, float downscale)
   {
       if(downscale <= 1)
           return;
